feat: implement no-show workflows, refactor booking user validation, and enhance webhook schema
