+++ apps/api/v2/src/modules/organizations/teams/memberships/organizations-teams-memberships.repository.ts
@@ -7,7 +7,7 @@ import { Injectable } from "@nestjs/common";
 
 @Injectable()
 export class OrganizationsTeamsMembershipsRepository {
-  constructor(private readonly dbRead: PrismaReadService, private readonly dbWrite: PrismaWriteService) {}
+  constructor(private readonly dbRead: PrismaReadService, private readonly dbWrite: PrismaWriteService) { }
 
   async findOrgTeamMembershipsPaginated(organizationId: number, teamId: number, skip: number, take: number) {
     return await this.dbRead.prisma.membership.findMany({
@@ -75,6 +75,13 @@ export class OrganizationsTeamsMembershipsRepository {
         teamId: teamId,
       },
       include: { user: { select: MembershipUserSelect } },
+    });
+  }
+
+  async findOrgTeam(teamId: number) {
+    return this.dbRead.prisma.team.findUnique({
+      where: { id: teamId },
+      include: { parent: true },
     });
   }
 }
