+++ packages/features/bookings/lib/service/RegularBookingService.ts
@@ -545,6 +545,7 @@ async function getEventOrganizationId({
 }: {
   eventType: {
     userId: number | null;
+    users: { id: number }[];
     team: {
       parentId: number | null;
     } | null;
@@ -563,10 +564,12 @@ async function getEventOrganizationId({
     return eventOrganizationId;
   }
 
-  if (eventType.userId) {
+  const organizerId = eventType.userId || (eventType.users.length > 0 ? eventType.users[0].id : null);
+
+  if (organizerId) {
     // TODO: Moving it to instance based access through DI in a followup
     const profile = await ProfileRepository.findFirstForUserId({
-      userId: eventType.userId,
+      userId: organizerId,
     });
     eventOrganizationId = profile?.organizationId ?? null;
     return eventOrganizationId;
@@ -623,9 +626,9 @@ async function handler(
     userId: userId ?? null,
     eventType: eventType
       ? {
-          seatsPerTimeSlot: eventType.seatsPerTimeSlot,
-          minimumRescheduleNotice: eventType.minimumRescheduleNotice ?? null,
-        }
+        seatsPerTimeSlot: eventType.seatsPerTimeSlot,
+        minimumRescheduleNotice: eventType.minimumRescheduleNotice ?? null,
+      }
       : null,
   });
 
@@ -809,10 +812,10 @@ async function handler(
         organizationId: eventOrganizationId,
         previousBooking: originalRescheduledBooking
           ? {
-              uid: originalRescheduledBooking.uid,
-              startTime: originalRescheduledBooking.startTime,
-              endTime: originalRescheduledBooking.endTime,
-            }
+            uid: originalRescheduledBooking.uid,
+            startTime: originalRescheduledBooking.startTime,
+            endTime: originalRescheduledBooking.endTime,
+          }
           : null,
       };
     }
@@ -1440,9 +1443,9 @@ async function handler(
   const { bookingLocation, conferenceCredentialId: eventTypeCredentialId } =
     organizerOrFirstDynamicGroupMemberDefaultLocationUrl
       ? {
-          bookingLocation: organizerOrFirstDynamicGroupMemberDefaultLocationUrl,
-          conferenceCredentialId: undefined,
-        }
+        bookingLocation: organizerOrFirstDynamicGroupMemberDefaultLocationUrl,
+        conferenceCredentialId: undefined,
+      }
       : getLocationValueForDB(locationBodyString, eventType.locations);
 
   // Use per-host credential if available, otherwise fall back to event type credential
@@ -1752,10 +1755,10 @@ async function handler(
       organizationId: eventOrganizationId,
       previousBooking: originalRescheduledBooking
         ? {
-            uid: originalRescheduledBooking.uid,
-            startTime: originalRescheduledBooking.startTime,
-            endTime: originalRescheduledBooking.endTime,
-          }
+          uid: originalRescheduledBooking.uid,
+          startTime: originalRescheduledBooking.startTime,
+          endTime: originalRescheduledBooking.endTime,
+        }
         : null,
     };
   }
@@ -1829,10 +1832,10 @@ async function handler(
         organizationId: eventOrganizationId,
         previousBooking: originalRescheduledBooking
           ? {
-              uid: originalRescheduledBooking.uid,
-              startTime: originalRescheduledBooking.startTime,
-              endTime: originalRescheduledBooking.endTime,
-            }
+            uid: originalRescheduledBooking.uid,
+            startTime: originalRescheduledBooking.startTime,
+            endTime: originalRescheduledBooking.endTime,
+          }
           : null,
       };
     } else {
@@ -2159,14 +2162,14 @@ async function handler(
 
     const updateManager = !skipCalendarSyncTaskCreation
       ? await eventManager.reschedule(
-          evt,
-          originalRescheduledBooking.uid,
-          undefined,
-          changedOrganizer,
-          previousHostDestinationCalendar,
-          isBookingRequestedReschedule,
-          skipDeleteEventsAndMeetings
-        )
+        evt,
+        originalRescheduledBooking.uid,
+        undefined,
+        changedOrganizer,
+        previousHostDestinationCalendar,
+        isBookingRequestedReschedule,
+        skipDeleteEventsAndMeetings
+      )
       : placeholderCreatedEvent;
 
     results = updateManager.results;
@@ -2461,17 +2464,17 @@ async function handler(
 
   const metadata = videoCallUrl
     ? {
-        videoCallUrl: getVideoCallUrlFromCalEvent(evt) || videoCallUrl,
-      }
+      videoCallUrl: getVideoCallUrlFromCalEvent(evt) || videoCallUrl,
+    }
     : undefined;
 
   // Query feature flags in parallel before firing booking events
   // TODO: We should support checkIfOrgHasFeatures - Bulk plus orgId check, that would be more efficient.
   const [isBookingEmailSmsTaskerEnabled, isBookingAuditEnabled] = orgId
     ? await Promise.all([
-        deps.featuresRepository.checkIfTeamHasFeature(orgId, "booking-email-sms-tasker"),
-        deps.featuresRepository.checkIfTeamHasFeature(orgId, "booking-audit"),
-      ])
+      deps.featuresRepository.checkIfTeamHasFeature(orgId, "booking-email-sms-tasker"),
+      deps.featuresRepository.checkIfTeamHasFeature(orgId, "booking-audit"),
+    ])
     : [false, false];
 
   await this.fireBookingEvents({
@@ -2561,9 +2564,9 @@ async function handler(
         ...eventType,
         metadata: eventType.metadata
           ? {
-              ...eventType.metadata,
-              apps: eventType.metadata?.apps as Prisma.JsonValue,
-            }
+            ...eventType.metadata,
+            apps: eventType.metadata?.apps as Prisma.JsonValue,
+          }
           : {},
       },
       paymentAppCredentials: eventTypePaymentAppCredential as IEventTypePaymentCredentialType,
@@ -2653,10 +2656,10 @@ async function handler(
       organizationId: eventOrganizationId,
       previousBooking: originalRescheduledBooking
         ? {
-            uid: originalRescheduledBooking.uid,
-            startTime: originalRescheduledBooking.startTime,
-            endTime: originalRescheduledBooking.endTime,
-          }
+          uid: originalRescheduledBooking.uid,
+          startTime: originalRescheduledBooking.startTime,
+          endTime: originalRescheduledBooking.endTime,
+        }
         : null,
     };
   }
@@ -2819,15 +2822,11 @@ async function handler(
   try {
     if (isConfirmedByDefault) {
       await scheduleNoShowTriggers({
-        booking: {
-          startTime: booking.startTime,
-          id: booking.id,
-          location: booking.location,
-          uid: booking.uid,
-        },
-        triggerForUser,
-        organizerUser: { id: organizerUser.id },
-        eventTypeId,
+        booking,
+        triggerForUser: true,
+        organizerUser: { id: booking.userId },
+        eventTypeId: booking.eventTypeId,
+        workflows,
         teamId,
         orgId,
         isDryRun,
@@ -2895,10 +2894,10 @@ async function handler(
     organizationId: eventOrganizationId,
     previousBooking: originalRescheduledBooking
       ? {
-          uid: originalRescheduledBooking.uid,
-          startTime: originalRescheduledBooking.startTime,
-          endTime: originalRescheduledBooking.endTime,
-        }
+        uid: originalRescheduledBooking.uid,
+        startTime: originalRescheduledBooking.startTime,
+        endTime: originalRescheduledBooking.endTime,
+      }
       : null,
   };
 }
@@ -2909,7 +2908,7 @@ async function handler(
  * We are open to renaming it to something more descriptive.
  */
 export class RegularBookingService implements IBookingService {
-  constructor(private readonly deps: IBookingServiceDependencies) {}
+  constructor(private readonly deps: IBookingServiceDependencies) { }
 
   async fireBookingEvents({
     booking,
@@ -2977,10 +2976,10 @@ export class RegularBookingService implements IBookingService {
         bookerName,
         rescheduledBy: rescheduledBy
           ? {
-              attendeeId: rescheduledByAttendeeId ?? null,
-              userUuid: rescheduledByUserUuid ?? null,
-              email: rescheduledBy,
-            }
+            attendeeId: rescheduledByAttendeeId ?? null,
+            userUuid: rescheduledByUserUuid ?? null,
+            email: rescheduledBy,
+          }
           : null,
         logger: tracingLogger,
       });
