+++ packages/features/tasker/tasks/sendWebook.ts
@@ -3,6 +3,20 @@ import { z } from "zod";
 import { WebhookVersion } from "@calcom/features/webhooks/lib/interface/IWebhookRepository";
 import sendPayload from "@calcom/features/webhooks/lib/sendPayload";
 
+const webhookDataSchema = z.union([
+  z.object({
+    id: z.number(),
+    uid: z.string(),
+    title: z.string(),
+    startTime: z.string(),
+    endTime: z.string(),
+    attendees: z.array(z.object({ email: z.string(), name: z.string() })),
+    organizer: z.object({ email: z.string(), name: z.string() }),
+  }),
+  // Add other payload types if needed, but for now this covers the main booking payload
+  z.any(),
+]);
+
 const sendWebhookPayloadSchema = z.object({
   secretKey: z.string().nullable(),
   triggerEvent: z.string(),
@@ -13,8 +27,7 @@ const sendWebhookPayloadSchema = z.object({
     payloadTemplate: z.string().nullable(),
     version: z.nativeEnum(WebhookVersion),
   }),
-  // TODO: Define the data schema
-  data: z.any(),
+  data: webhookDataSchema,
 });
 
 export async function sendWebhook(payload: string): Promise<void> {
